requirements.txt
accelerate==0.31.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blobfile==2.1.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.6.1
datasets==2.14.6
dill==0.3.7
distro==1.8.0
exceptiongroup==1.1.3
fairscale==0.4.13
fastapi==0.103.2
ffmpy==0.3.1
filelock==3.13.0
fonttools==4.43.1
frozenlist==1.4.0
fschat==0.2.36
fsspec==2023.10.0
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-search-results==2.4.2
googleapis-common-protos==1.61.0
gradio==3.50.2
gradio_client==0.6.1
greenlet==3.0.1
h11==0.14.0
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.317
langsmith==0.0.53
litellm==0.12.5
loguru==0.7.2
lxml==4.9.4
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.1
nh3==0.2.17
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.3.5
orjson==3.9.10
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
pathlib==1.0.1
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
protobuf==4.24.4
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.20.0
pydantic==1.10.9
pydub==0.25.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.10.1
selenium==4.14.0
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
serpapi==0.1.4
shortuuid==1.0.13
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.22
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
tenacity==8.2.3
text2vec==1.2.9
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
toolz==0.12.0
torch==2.1.0
torch-tb-profiler==0.4.3
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.1
trio==0.22.2
trio-websocket==0.11.1
triton==2.1.0
types-colorama==0.4.15.12
types-requests==2.31.0.10
typing-inspect==0.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.23.2
wavedrom==2.0.3.post3
websockets==11.0.3
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
